# Simplify memoryengine