Instead of relying on a goroutine, the memoryengine now
adds reads from the underlying btree in batch and then
iterates over the cached items.
This makes the memoryengine smaller and much faster.

```
name                            old time/op    new time/op    delta
MemoryEngineStoreScan/00001-12    4.04µs ± 0%    0.23µs ± 0%  -94.21%  (p=1.000 n=1+1)
MemoryEngineStoreScan/00010-12    6.86µs ± 0%    0.37µs ± 0%  -94.60%  (p=1.000 n=1+1)
MemoryEngineStoreScan/00100-12    37.3µs ± 0%     1.9µs ± 0%  -94.96%  (p=1.000 n=1+1)
MemoryEngineStoreScan/01000-12     335µs ± 0%      21µs ± 0%  -93.68%  (p=1.000 n=1+1)
MemoryEngineStoreScan/10000-12    3.28ms ± 0%    0.24ms ± 0%  -92.77%  (p=1.000 n=1+1)

name                            old alloc/op   new alloc/op   delta
MemoryEngineStoreScan/00001-12      656B ± 0%      656B ± 0%     ~     (all equal)
MemoryEngineStoreScan/00010-12      656B ± 0%      656B ± 0%     ~     (all equal)
MemoryEngineStoreScan/00100-12      656B ± 0%      656B ± 0%     ~     (all equal)
MemoryEngineStoreScan/01000-12      656B ± 0%      656B ± 0%     ~     (all equal)
MemoryEngineStoreScan/10000-12      656B ± 0%      656B ± 0%     ~     (all equal)

name                            old allocs/op  new allocs/op  delta
MemoryEngineStoreScan/00001-12      8.00 ± 0%      2.00 ± 0%  -75.00%  (p=1.000 n=1+1)
MemoryEngineStoreScan/00010-12      8.00 ± 0%      2.00 ± 0%  -75.00%  (p=1.000 n=1+1)
MemoryEngineStoreScan/00100-12      8.00 ± 0%      2.00 ± 0%  -75.00%  (p=1.000 n=1+1)
MemoryEngineStoreScan/01000-12      8.00 ± 0%      2.00 ± 0%  -75.00%  (p=1.000 n=1+1)
MemoryEngineStoreScan/10000-12      8.00 ± 0%      2.00 ± 0%  -75.00%  (p=1.000 n=1+1)
```